add auth to deleteProduct, update utilities module

+++ src/functions/deleteProduct/handler.ts
@@ -2,7 +2,7 @@ import 'source-map-support/register';
 
 import Product from '@dbModel/tables/product';
 import deleteProduct from './function';
-import { ValidatedEventAPIGatewayProxyEvent, middyfy, Response } from 'utilities-techsweave';
+import { ValidatedEventAPIGatewayProxyEvent, middyfy, Response, AuthenticatedUser } from 'utilities-techsweave';
 
 import { StatusCodes } from 'http-status-codes';
 /*
@@ -12,6 +12,15 @@ import { StatusCodes } from 'http-status-codes';
 const deleteProductHandler: ValidatedEventAPIGatewayProxyEvent<void> = async (event) => {
     let res: Response<Product>;
     try {
+
+        const user: AuthenticatedUser = await AuthenticatedUser.fromToken(event.headers?.Authorization);
+        if (!(await user.isVendor(process.env.USER_POOL_ID))) {
+            throw {
+                name: 'userNotAllowed',
+                message: 'You must be a vendor to create a product'
+            };
+        }
+
         res = Response.fromData<Product>(
             await deleteProduct(event.pathParameters.id),
             StatusCodes.OK);
